feat(csharp/src/Drivers/Databricks): Fixes to heartbeat polling (#2851)
`DatabricksOperationStatusPoller` was interleaving calls with
`CloudFetchResultFetcher`. Since they both share the underlying
transport and protocol to perform this operation (and `TTransport` is
not thread-safe), it resulted in race conditions between status polling
and result fetching.

This PR fixes this by adding a ThreadSafeClient to classes sharing a
connection/protocol/transport. This PR should not affect Impala's client


### Testing
Ran CloudFetchE2Etests, passes on this branch (fails on main)