feat(csharp/test/Drivers/Databricks): Enable RunAsync option in TExecuteStatementReq (#3171)
## Motivation

Databricks adds `RunAsync` (default is false) option for all the thrift
operations, when it is set as `true`, the operation runs async at the
backend and the status can be polled by calling getStatus, it helps
Databricks backend to better manage capacity and load on client
requests. Furthermore, it can avoid the unnecessary retry on thrift
operation (e.g. TExecuteStatementReq) when the warehouse is stopped or
unavailable, the backend will returns 503 error when `RunAsync` is false
and client has to retry on this thrift operation till the warehouse is
up, this generates lots of queries with 503 errors in the Databricks
query history and consume more resources. When `RunAsync=true`, server
will return 200 with query state `PENDING`, the client will poll the
status till the warehouse is up, this only generates one query in the
query history.

## Change
- Add a connection parameter `adbc.databricks.enable_run_async_thrift`,
default is `false` (it will be changed to `true` later)
- Set `RunAsync` of `TExecuteStatementReq` with above connection
parameter (`RunAsyncInThrift`) in
`DatabricksStatement:SetStatementProperties`
- Fix a bug in `BaseDatabricksReader` by adding `null` check on
`statement.DirectResults.ResultSet`
- Fix the case `TestVarcharExceptionDataDatabricks` in
`StringValuesTest`: when error is in the `DirectResult.status` (case for
`RunAsync=true`), it throw the error with `DisplayMessage` (see
[here](https://github.com/apache/arrow-adbc/blob/main/csharp/src/Drivers/Apache/Hive2/HiveServer2Statement.cs#L316))
instead of the `Message` and `DisplayMessage` does not include some
internal error info such as error exception class at the backend.


## Test
- Run all the E2E tests under `csharp/test/Drivers/Databricks/E2E` when
`Connection.RunAsyncInThrift` is both on and off